serviceChanged: Fix session affinity, type
When comparing services to determine whether an update is required, treat
implicit and explicit default values as equal.

Before this commit, the update logic would keep trying to revert the
default value that the API set for the service's session affinity or type.

* pkg/operator/controller/controller_dns_service.go (serviceChanged): Use
the new cmpServiceAffinity and cmpServiceType functions so that two
services are considered equal if the only difference between them is that
one does not specify the session affinity or type and the other service has
the default values.
(cmpServiceAffinity, cmpServiceType): New function.
* pkg/operator/controller/controller_dns_service_test.go:
(TestDNSServiceChanged): Verify that serviceChanged returns false if the
only mutation is that the session affinity or type has been updated from
empty to the default value.