daemonsetConfigChanged: Fix for volume default mode
When comparing daemonsets to determine whether an update is required, treat
implicit and explicit default values as equal.

Before this commit, the update logic would keep trying to revert the
default value that the API set for a volume's default mode.

* pkg/operator/controller/controller_dns_daemonset.go
(daemonsetConfigChanged): Use the new cmpConfigMapVolumeSource and
cmpSecretVolumeSource functions so that two daemonsets are considered equal
if the only difference between them is that one does not specify a default
mode value and the other daemonset specifies the default value.
(cmpConfigMapVolumeSource, cmpSecretVolumeSource): New functions.
* pkg/operator/controller/controller_dns_daemonset_test.go:
(TestDaemonsetConfigChanged): Verify that daemonsetConfigChanged returns
false if the only mutation is that a volume's default mode value has been
updated from empty to the default value.