Add node resolver daemonset
Split the node resolver out of the DNS daemonset into a separate daemonset.

The node resolver daemonset needs to run on every node and therefore
must tolerate all taints.  Because the node resolver runs in a separate
daemonset, the blanket toleration can be removed from the DNS daemonset.

* assets/dns/daemonset.yaml: Remove the "dns-node-resolver" container,
"hosts-file" volume, and blanket toleration.
* assets/dns/update-node-resolver.sh: New file.
* pkg/manifests/bindata.go: Regenerate.
* pkg/manifests/manifests.go (nodeResolverScriptAsset): New const with the
path to the node resolver script asset.
(NodeResolverScript): New function that returns the node resolver script.
* pkg/manifests/manifests_test.go (TestManifests): Verify
NodeResolverScript.
* pkg/operator/controller/controller.go (ensureDNS): Delete the clusterIP
and clusterDomain arguments to ensureDNSDaemonSet.  Call
ensureNodeResolverDaemonset after ensureDNSDaemonSet.
* pkg/operator/controller/controller_dns_daemonset.go (ensureDNSDaemonSet):
Delete the clusterIP and clusterDomain parameters, and delete the
clusterIP, clusterDomain, and openshiftCLIImage arguments in the call to
desiredDNSDaemonSet.
(desiredDNSDaemonSet): Delete the clusterIP, clusterDomain, and
openshiftCLIImage parameters and logic to configure the node resolver
container.
(daemonsetConfigChanged): Delete logic for the node resolver container.
* pkg/operator/controller/controller_dns_daemonset_test.go
(TestDesiredDNSDaemonset, TestDaemonsetConfigChanged): Delete logic related
to the node resolver container.
* pkg/operator/controller/controller_dns_node_resolver_daemonset.go
(services): New const with the list of service names for which entries
should be added to /etc/hosts.
(nodeResolverScript): New variable with the shell script that updates
/etc/hosts.
(ensureNodeResolverDaemonset): Add logic to create or update the node
resolver daemonset as needed, using the new createNodeResolverDaemonSet and
updateNodeResolverDaemonSet methods.
(desiredNodeResolverDaemonSet): Return true and a daemonset.
(createNodeResolverDaemonSet): New method.
(updateNodeResolverDaemonSet): New method that updates the node resolver
daemonset using the new nodeResolverDaemonSetConfigChanged function.
(nodeResolverDaemonSetConfigChanged): New function that checks if the
current daemonset needs to be updated, and if so, return true and an
updated version of the daemonset.
(TestDesiredNodeResolverDaemonset): Verify that
desiredNodeResolverDaemonSet returns true and a valid daemonset.
(TestNodeResolverDaemonsetConfigChanged): New test that verifies that
nodeResolverDaemonSetConfigChanged correctly detects changes to the node
resolver daemonset.
* pkg/operator/controller/names.go (nodeResolverDaemonSetLabelName): New
const with the name of the label that identifies the node resolver
daemonset.
(NodeResolverDaemonSetPodSelector): New function that returns the label
selector for node resolver pods.