serviceChanged: Fix for clusterIPs defaulting
Ignore the clusterIPs field when comparing services to determine whether an
update is required.

Before this commit, the update logic would keep trying to revert the value
that the API set for the service's clusterIPs field.

The clusterIPs field is new in Kubernetes 1.20 (OpenShift 4.7).

This commit is related to bug 1936022.

https://bugzilla.redhat.com/show_bug.cgi?id=1936022

* pkg/operator/controller/controller_dns_service.go
(serviceChanged): Ignore the services' clusterIPs field.
* pkg/operator/controller/controller_dns_service_test.go
(TestDNSServiceChanged): Verify that serviceChanged returns false if the
only mutation is that the clusterIPs field's value has changed.